refactor: clean up the code to prepare the site-wide experiment (#471)
