+++ 404.html
@@ -49,6 +49,7 @@
     window.clarity ||= function() {
       (window.clarity.q ||= []).push(arguments)
     }
+    window.PushlySDK ||= [];
     function gtag() { window.dataLayer.push(arguments); }
     gtag('consent', 'default', {
       'ad_storage': 'denied',
